Use pretty_assertions crate for tests.
This simplifies the FakePrinter and Validator logic and gives much
nicer assertion errors when tests fail.

This imposes a new requirement for tests to fully specify the printed
tasks, but this is probably for the better.
