+++ printing/src/lib.rs
@@ -6,7 +6,6 @@ pub mod printable_task;
 pub mod printable_warning;
 pub mod scripting_todo_printer;
 pub mod simple_todo_printer;
-pub mod testing;
 pub mod todo_printer;
 
 pub use self::brief_printable_task::*;
@@ -16,7 +15,6 @@ pub use self::printable_task::*;
 pub use self::printable_warning::*;
 pub use self::scripting_todo_printer::*;
 pub use self::simple_todo_printer::*;
-pub use self::testing::*;
 pub use self::todo_printer::*;
 
 #[cfg(test)]
@@ -24,13 +22,9 @@ mod printable_error_test;
 #[cfg(test)]
 mod printable_info_test;
 #[cfg(test)]
-mod printable_task_test;
-#[cfg(test)]
 mod printable_warning_test;
 #[cfg(test)]
 mod simple_todo_printer_test;
-#[cfg(test)]
-mod testing_test;
 
 #[derive(Default)]
 pub struct PrintableAppSuccess<'list> {
